requirements.txt
flask>=3.0
flask-socketio>=5.3
pyyaml>=6.0
watchdog>=4.0
eventlet>=0.35
